# Speed up ST_RemoveRepeatedPoints