--- a/postgis/lwgeom_functions_basic.c
+++ b/postgis/lwgeom_functions_basic.c
@@ -2748,11 +2748,10 @@ Datum ST_RemoveRepeatedPoints(PG_FUNCTION_ARGS);
 PG_FUNCTION_INFO_V1(ST_RemoveRepeatedPoints);
 Datum ST_RemoveRepeatedPoints(PG_FUNCTION_ARGS)
 {
-	GSERIALIZED *g_in = PG_GETARG_GSERIALIZED_P(0);
-	int type = gserialized_get_type(g_in);
+	GSERIALIZED *g_in = PG_GETARG_GSERIALIZED_P_COPY(0);
+	uint32_t type = gserialized_get_type(g_in);
 	GSERIALIZED *g_out;
 	LWGEOM *lwgeom_in = NULL;
-	LWGEOM *lwgeom_out = NULL;
 	double tolerance = 0.0;
 
 	/* Don't even start to think about points */
@@ -2763,22 +2762,15 @@ Datum ST_RemoveRepeatedPoints(PG_FUNCTION_ARGS)
 		tolerance = PG_GETARG_FLOAT8(1);
 
 	lwgeom_in = lwgeom_from_gserialized(g_in);
-	lwgeom_out = lwgeom_remove_repeated_points(lwgeom_in, tolerance);
+	lwgeom_remove_repeated_points_in_place(lwgeom_in, tolerance);
 
 	/* COMPUTE_BBOX TAINTING */
 	if (lwgeom_in->bbox)
-		lwgeom_refresh_bbox(lwgeom_out);
+		lwgeom_refresh_bbox(lwgeom_in);
 
-	g_out = geometry_serialize(lwgeom_out);
+	g_out = geometry_serialize(lwgeom_in);
 
-	if (lwgeom_out != lwgeom_in)
-	{
-		lwgeom_free(lwgeom_out);
-	}
-
-	lwgeom_free(lwgeom_in);
-
-	PG_FREE_IF_COPY(g_in, 0);
+	pfree(g_in);
 	PG_RETURN_POINTER(g_out);
 }