--- a/postgis/lwgeom_functions_basic.c
+++ b/postgis/lwgeom_functions_basic.c
@@ -2753,6 +2753,7 @@ Datum ST_RemoveRepeatedPoints(PG_FUNCTION_ARGS)
 	GSERIALIZED *g_out;
 	LWGEOM *lwgeom_in = NULL;
 	double tolerance = 0.0;
+	int modified = LW_FALSE;
 
 	/* Don't even start to think about points */
 	if (type == POINTTYPE)
@@ -2762,11 +2763,12 @@ Datum ST_RemoveRepeatedPoints(PG_FUNCTION_ARGS)
 		tolerance = PG_GETARG_FLOAT8(1);
 
 	lwgeom_in = lwgeom_from_gserialized(g_in);
-	lwgeom_remove_repeated_points_in_place(lwgeom_in, tolerance);
-
-	/* COMPUTE_BBOX TAINTING */
-	if (lwgeom_in->bbox)
-		lwgeom_refresh_bbox(lwgeom_in);
+	modified = lwgeom_remove_repeated_points_in_place(lwgeom_in, tolerance);
+	if (!modified)
+	{
+		/* Since there were no changes, we can return the input to avoid the serialization */
+		PG_RETURN_POINTER(g_in);
+	}
 
 	g_out = geometry_serialize(lwgeom_in);