### First query

The table contains 1083175 multipolygons with an average of 9.01 points per geometry.

```sql
explain analyze Select ST_RemoveRepeatedPoints(the_geom_webmercator, 1) from public.benchmark_af85af967525e69373b7b8678f3bd409a8e78f447671344b7535d;
```

Trunk:
```
duration: 20 s
number of transactions actually processed: 66
latency average = 306.126 ms
```

With this PR:
```
number of transactions actually processed: 77
latency average = 263.052 ms
```


### Second query

13 multipolygons with an average of 260843.07 points per geometry.

```sql
explain analyze Select ST_RemoveRepeatedPoints(the_geom_webmercator, 1) from public.benchmark_4c7214d90a79aa6760367a084a4d4a2f61fbe1c6cc4f7f9e76020;
```

Trunk:
```
number of transactions actually processed: 241
latency average = 83.137 ms
```

With this PR:
```
number of transactions actually processed: 375
latency average = 53.409 ms
```

### Explanation

- PG facing (`ST_RemoveRepeatedPoints`)
  - Use `PG_GETARG_GSERIALIZED_P_COPY` and `lwgeom_remove_repeated_points_in_place`: In some cases (like the second query) this avoids an extra geometry clone (one to copy the gserialized and the second one done in lwgeom_remove_repeated_points).
  - Remove the calls to `lwgeom_free`. It will be freed by PG automatically in a faster way. I've kept the call to free the GSERIALIZED copy because the comment in PGs code doesn't give me enough peace of mind (https://github.com/postgres/postgres/blob/bde7493d10898831100a0c6c233a5f3030bfcecd/src/include/fmgr.h#L251-L253)

- liblwgeom (`ptarray_remove_repeated_points_in_place`):
  - Keep a reference to where we want to write and use memcpy instead of `ptarray_copy_point`. With this we avoid calculating the position every time and also the size of the point (which won't change).

  ~~- Move the last_point special case outside the loop. This doesn't have a really big impact since the CPU is pretty good at predicting those branches.~~