@Algunenano 2019-08-28
> I'm not sure this is a good behavior to set on a function-by-function basis. Can you set lwfree to a no-op if you want to let PG handle cleanup?

AFAIK, yes you can with some exceptions. In fact I created an issue related to this: https://trac.osgeo.org/postgis/ticket/4264 but I haven't had time to look into this in detail.

Basically the idea is that PG resets the memory context after every transaction, so you don't need to delete intermediate objects unless they are going to amount to a lot (big allocations, loops, recursive functions) or you are not using PG allocation (GEOS objects). 

I believe this is common practise in Postgres codebase, but I can't find any comment to reference it now, just examples:
- The ASCII functions: https://github.com/postgres/postgres/blob/8255c7a5eeba8f1a38b7a431c04909bde4f5e67d/src/backend/utils/adt/ascii.c#L120

And counter examples:
- The varlena functions (Used for sorting thus the garbage is kept longer?): https://github.com/postgres/postgres/blob/8255c7a5eeba8f1a38b7a431c04909bde4f5e67d/src/backend/utils/adt/varlena.c#L2135