@Algunenano 2019-08-28
I totally get what you mean and, ideally, I'd like to just make lwfree a no-op; but to do that I'd need to review the hundreds of calls to lwalloc (and it's callers) to make their deletion can be deferred until the end of the transaction.

One thing to note is that I removed the call to lwfree in the extension code (`postgis/`), not in the library (`liblwgeom`) which should still be memory safe, it's just that I've chosen not to manually delete the memory I've gotten from the library because I know it's going to be recovered as soon as Postgres removes the memory context (which is done immediately after and it's faster as it doesn't need to traverse structures).

There are other places (also in the extension code) that also make use of this, for example `mvt_geom` doesn't care if you either modify or create a new geometry in `mvt_clip_and_validate`; it will return just that and leaves Postgres the task of deleting everything at the end. Looking through the code I see there are other functions (`ST_DistanceRectTree`, `geography_point_outside`) functions don't call `lwgeom_free` before returning but most of them do, although not always in error handling.