@Algunenano 2019-08-29
I realized that changing the gserialized read from the normal (PG_GETARG_GSERIALIZED_P) to a copy (PG_GETARG_GSERIALIZED_P_COPY) might have a impact on points so I decided to test it too with a 1M rows dataset:

```
explain analyze Select ST_RemoveRepeatedPoints(the_geom_webmercator, 1) from public.yellow_1m;
```

Trunk:
```
number of transactions actually processed: 144
latency average = 139.095 ms
```

This PR:
```
number of transactions actually processed: 143
latency average = 140.610 ms
```

So there might be a slight penalty for that case (a no-op) but it's not clear.