@Algunenano 2019-09-13
After rebasing on top of trunk and adding a shortcut to optimize the no-op case (return the input when no points were removed):
* Trunk
- Tolerance: 0.00001 (no-op): ~220 its / 30 secs
- Tolerance: 1 (~1% points removed): ~230 its / 30 secs
- Tolerance: 1000000 (~98% of points removed): ~360 its / 30 secs

* This PR
- Tolerance: 0.00001 (no-op): ~615 its / 30 secs --> **2.8x as fast**
- Tolerance: 1 (~1% points removed): ~350 its / 30 secs --> **1.5x as fast**
- Tolerance: 1000000 (~98% of points removed): ~720 its / 30 secs --> **2x as fast**

Dataset: 13 multipolygons with an average of 260843.07 points per geometry