@dr-jts 2019-09-13
> I totally get what you mean and, ideally, I'd like to just make lwfree a no-op; but to do that I'd need to review the hundreds of calls to lwalloc (and it's callers) to make their deletion can be deferred until the end of the transaction.

How about making a second function `lwfree_noop` (or something) which is a no-op, and has a comment indicating that this only works if running in the context of a memory pool manager?

This captures the fact that memory has been allocated and would need to be freed if not running under the security of Postgres?