@dr-jts 2019-09-13
> > How about making a second function lwfree_noop (or something) which is a no-op, and has a comment indicating that this only works if running in the context of a memory pool manager?
> 
> That's a nice way to distinguish between them, but again it requires revisiting all lwfree calls to distinguish between those where it's safe to not free them under Postgresql context and those where it isn't. 

> In the current case in particular, the lwfree that was removed is always under postgres context (it's not inside the lwgeom library) so removing it it's safe.

I wasn't thinking that work would need to be done right away.  Just that it would establish a new pattern for future cleanup, and make this particular case visible. But if this `lwfree` isn't needed ever, then it makes sense to just remove it.