@strk 2019-09-16
On Fri, Sep 13, 2019 at 09:50:57AM -0700, Martin Davis wrote:
> > I totally get what you mean and, ideally, I'd like to just make lwfree a no-op; but to do that I'd need to review the hundreds of calls to lwalloc (and it's callers) to make their deletion can be deferred until the end of the transaction.
> 
> How about making a second function `lwfree_noop` (or something) which is a no-op, and has a comment indicating that this only works if running in the context of a memory pool manager?
> 
> This captures the fact that memory has been allocated and would need to be freed if not running under the security of Postgres?

`lwfree` is supposed to be definable by user. PostgreSQL user defines it
to `pfree`, while cunit user defines it to `free`.

Please beware of raises in memory usage when postponing release, test
with big tables.